# OpenAPI/OpenAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SEC-API4CEP",
    "description": "Secure RESTful API for runtime management of a CEP engine",
    "contact": {
      "name": "Manuel Cano-Crespo",
      "email": "manuel.cano@uca.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://choosealicense.com/licenses/mit/"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost:443",
      "description": "Server URL in Development environment"
    }
  ],
...
  "paths": {
...
    "/event-type/{id}/undeploy": {
      "put": {
        "tags": [
          "1. Event Types"
        ],
        "summary": "Undeploys an existing event type identified by its ID",
        "description": "Undeploys an existing event type identified by its ID and sets it as not deployed in the database",
        "operationId": "undeploy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Bad credentials. You must be properly authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
...
    "/event-pattern/{id}": {
      "get": {
        "tags": [
          "2. Event Patterns"
        ],
        "summary": "Retrieves the event pattern with the provided ID",
        "description": "Retrieves the event pattern with the provided ID",
        "operationId": "getEventPatternById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event Pattern successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {

                  }
                }
              }
            }
          },
          "404": {
            "description": "Event Pattern not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {

                  }
                }
              }
            }
          },
          "403": {
            "description": "Bad credentials. You must be properly authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {

                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Authentication": []
          }
        ]
      },
      "put": {
        "tags": [
          "2. Event Patterns"
        ],
        "summary": "Updates the event pattern with the provided ID",
        "description": "Updates an existing pattern identified by its ID in the database and in the CEP engine if necessary",
        "operationId": "updateEventPattern",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPattern"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Event pattern succesfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Event pattern has not been updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bad credentials. You must be properly authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Authentication": []
          }
        ]
      },
      "delete": {
        "tags": [
          "2. Event Patterns"
        ],
        "summary": "Deletes the event pattern with the provided ID",
        "description": "Deletes the event pattern with the provided ID",
        "operationId": "deleteEventPattern",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Bad credentials. You must be properly authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
...
    "/event-pattern/{id}/undeploy": {
      "put": {
        "tags": [
          "2. Event Patterns"
        ],
        "summary": "Undeploys an existing pattern identified by its ID",
        "description": "Undeploys an existing pattern identified by its ID and sets it as not deployed in the database",
        "operationId": "undeploy_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Bad credentials. You must be properly authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
...
    "/event-pattern/{id}/deploy": {
      "put": {
        "tags": [
          "2. Event Patterns"
        ],
        "summary": "Deploys an existing pattern identified by its ID",
        "description": "Deploys an existing pattern identified by its ID and sets it as deployed in the database",
        "operationId": "deploy_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Bad credentials. You must be properly authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
...
    "/event-pattern": {
      "get": {
        "tags": [
          "2. Event Patterns"
        ],
        "summary": "Retrieves all existing event patterns",
        "description": "Retrieves all existing event patterns",
        "operationId": "getAllEventPatterns",
        "responses": {
          "403": {
            "description": "Bad credentials. You must be properly authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventPattern"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Authentication": []
          }
        ]
      },
      "post": {
        "tags": [
          "2. Event Patterns"
        ],
        "summary": "Registers a new event pattern",
        "description": "Registers a new event pattern",
        "operationId": "createEventPattern",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPattern"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Event pattern successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {

                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {

                  }
                }
              }
            }
          },
          "403": {
            "description": "Bad credentials. You must be properly authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {

                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
...
    "/event-type/{name}/byName": {
      "get": {
        "tags": [
          "1. Event Types"
        ],
        "summary": "Retrieves the event type with the provided name",
        "description": "Retrieves the event type with the provided name",
        "operationId": "findByName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Bad credentials. You must be properly authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {

                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
    "/event-pattern/{name}/byName": {
      "get": {
        "tags": [
          "2. Event Patterns"
        ],
        "summary": "Retrieves the event pattern with the provided name",
        "description": "Retrieves the event pattern with the provided name",
        "operationId": "findByName_1",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Bad credentials. You must be properly authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {

                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer Authentication": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventPattern": {
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "name": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "readyToDeploy": {
            "type": "boolean",
            "default": false
          },
          "deployed": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}